Renamed call_consensus (#49)
* Renamed call_consensus

* Updated about-strings

* Updated doc-strings

* Updated about-strings
