+++ src/bam/collapse_reads_to_fragments/calc_consensus.rs
