+++ src/bam/collapse_reads_to_fragments/mod.rs
