+++ src/bam/mod.rs
@@ -1,3 +1,3 @@
 //! Tools that work on BAM files
-pub mod call_consensus_reads;
+pub mod collapse_reads_to_fragments;
 pub mod depth;
