+++ src/fastq/collapse_reads_to_fragments/pipeline.rs
