+++ src/fastq/mod.rs
@@ -1,4 +1,4 @@
 //! Tools that work on FASTQ files
-pub mod call_consensus_reads;
+pub mod collapse_reads_to_fragments;
 pub mod filter;
 pub mod split;
