+++ tests/lib.rs
@@ -154,11 +154,11 @@ fn vcf_baf() {
 }
 
 #[test]
-fn test_call_consensus_reads_two_cluster() {
+fn test_collapse_reads_to_fragments_two_cluster() {
     assert!(
         Command::new("bash")
                 .arg("-c")
-                .arg("target/debug/rbt call-consensus-reads fastq --umi-len 3 -u --max-umi-dist 0 --max-seq-dist 2 tests/test-consensus.fastq tests/test-consensus.fastq /tmp/test-consensus.1.fastq /tmp/test-consensus.2.fastq")
+                .arg("target/debug/rbt collapse-reads-to-fragments fastq --umi-len 3 -u --max-umi-dist 0 --max-seq-dist 2 tests/test-consensus.fastq tests/test-consensus.fastq /tmp/test-consensus.1.fastq /tmp/test-consensus.2.fastq")
             .spawn().unwrap().wait().unwrap().success());
     compare_fastq(
         "/tmp/test-consensus.1.fastq",
@@ -171,11 +171,11 @@ fn test_call_consensus_reads_two_cluster() {
 }
 
 #[test]
-fn test_call_consensus_reads_single_cluster() {
+fn test_collapse_reads_to_fragments_single_cluster() {
     assert!(
         Command::new("bash")
             .arg("-c")
-            .arg("target/debug/rbt call-consensus-reads fastq --umi-len 3 -u --max-umi-dist 2 --max-seq-dist 2 tests/test-consensus.fastq tests/test-consensus.fastq /tmp/test-consensus_single.1.fastq /tmp/test-consensus_single.2.fastq")
+            .arg("target/debug/rbt collapse-reads-to-fragments fastq --umi-len 3 -u --max-umi-dist 2 --max-seq-dist 2 tests/test-consensus.fastq tests/test-consensus.fastq /tmp/test-consensus_single.1.fastq /tmp/test-consensus_single.2.fastq")
             .spawn().unwrap().wait().unwrap().success());
     compare_fastq(
         "/tmp/test-consensus_single.1.fastq",
@@ -188,11 +188,11 @@ fn test_call_consensus_reads_single_cluster() {
 }
 
 #[test]
-fn test_call_overlapping_consensus_reads() {
+fn test_collapse_reads_to_fragments_reads() {
     assert!(
         Command::new("bash")
             .arg("-c")
-            .arg("target/debug/rbt call-consensus-reads fastq --umi-len 10 --max-umi-dist 0 --max-seq-dist 8 --insert-size 450 --std-dev 50  tests/overlapping-consensus.1.fastq tests/overlapping-consensus.2.fastq /tmp/test_overlapping-consensus.1.fastq /tmp/test_overlapping-consensus.2.fastq /tmp/test_overlapping-consensus.3.fastq")
+            .arg("target/debug/rbt collapse-reads-to-fragments fastq --umi-len 10 --max-umi-dist 0 --max-seq-dist 8 --insert-size 450 --std-dev 50  tests/overlapping-consensus.1.fastq tests/overlapping-consensus.2.fastq /tmp/test_overlapping-consensus.1.fastq /tmp/test_overlapping-consensus.2.fastq /tmp/test_overlapping-consensus.3.fastq")
             .spawn().unwrap().wait().unwrap().success());
     compare_fastq(
         "/tmp/test_overlapping-consensus.1.fastq",
@@ -209,11 +209,11 @@ fn test_call_overlapping_consensus_reads() {
 }
 
 #[test]
-fn test_call_consensus_from_bam() {
+fn test_collapse_reads_to_fragments_from_bam() {
     assert!(
     Command::new("bash")
         .arg("-c")
-        .arg("target/debug/rbt call-consensus-reads bam --max-seq-dist 8 tests/overlapping_consensus_marked.bam /tmp/overlapping_consensus_marked.bam")
+        .arg("target/debug/rbt collapse-reads-to-fragments bam --max-seq-dist 8 tests/overlapping_consensus_marked.bam /tmp/overlapping_consensus_marked.bam")
         .spawn().unwrap().wait().unwrap().success());
     compare_bam(
         "/tmp/overlapping_consensus_marked.bam",
